CEDS-1287 Move the responsibility of RegEx to the individual loaders to ensure we handle different formats
