+++ conf/office-of-exits.csv
@@ -1,0 +1,132 @@
+Description,Code
+Aberdeen,GB000434
+Aberdeen Airport,GB000411
+Appledore c/o National Simps Team,GB005010
+Avonmouth 1,GB000033
+Barrow in Furness,GB003010
+Barry Dock,GB000125
+Belfast EPU,GB000142
+Belfast International Airport,GB000014
+Berwick upon Tweed,GB003020
+Biggin Hill,GB003290
+Birmingham Airport,GB000011
+Blackpool International Airport,GB003030
+Boston,GB000244
+Bournemouth (Hurn) Airport,GB003060
+Bridgwater c/o National Simps Team,GB005020
+Brize Norton,GB003070
+Cambridge Airport,GB003080
+Cardiff,GB000328
+Cardiff International Airport,GB000122
+Central Community Transit Office,GB000001
+Chatham,GB000480
+Coventry International Hub,GB000500
+Croydon Regional Centre,GB005211
+Dagenham c/o National Simps Team,GB005030
+Dartford c/o National Simps Team,GB005040
+Dorchester House,GBREX001
+Dover (OTS) Freight Clearance,GB000040
+Dover/Folkestone Eurotunnel Freight,GB000060
+Dundee (Sea),GB003100
+Durham Tees Valley (Teeside Apt),GB003180
+Edinburgh Airport,GB000128
+Exeter Airport,GB000297
+Falmouth,GB000108
+Farnborough Airport,GB003110
+Felixstowe,GB000051
+Fishguard,GB005200
+Folkestone,GB000041
+Fords Jetty,GB000921
+George Best (Belfast City),GB000140
+Glasgow,GB004098
+Glasgow Docks,GB003040
+Gloucester (Staverton) Airport,GB003120
+Grangemouth,GB000059
+Great Yarmouth,GB000054
+Greenock c/o Simps Team,GB005050
+Grimsby,GB000073
+Guernsey Customs & Excise,GB000461
+Hartlepool,GB003140
+Harwich,GB000052
+Heysham,GB005210
+Holyhead,GB000124
+Hull,GB000072
+Humberside International Airport,GB003130
+HWDC Langley,GB000501
+Immingham,GB000074
+Inverness,GB003150
+Ipswich Port,GB000053
+Isle Of Man Customs & Excise,GB000450
+Jersey Customs & Immigration,GB000465
+Kemble Airport,GB003160
+Killingholme,GB005170
+King's Lynn,GB000055
+Kirkwall (Orkney Islands),GB003170
+Langstone c/o Simps Team,GB005060
+Leeds Bradford Airport,GB000087
+Lerwick,GB001016
+Liverpool John Lennon Airport,GB003190
+Liverpool Seaforth S2 Berth,GB000080
+London Ashford (Lydd) Airport,GB003200
+London City Airport,GB000130
+London Gateway Port,GB000170
+London Gatwick Airport Cargo Dist.,GB000085
+London Heathrow Airport (Baggage),GB000120
+London Heathrow Cargo,GB000084
+London c/o Simps Team,GB005070
+Londonderry,GB003210
+Luton Airport Fct,GB000102
+Manchester,GB000191
+Manchester Docks,GB003220
+Middlesbrough c/o Simps Team,GB005080
+Montrose,GB005180
+Neath c/o Simps Team,GB005090
+Newcastle Airport,GB000098
+Newcastle-Upon-Tyne,GB000218
+Newhaven,GB003250
+Newport,GB000049
+Northolt Airport,GB000119
+Norwich Airport,GB000058
+Oxford (Kidlington) Airport,GB003230
+Pembroke Dock,GB000126
+Penrhyn c/o Simps Team,GB005100
+Peter Bennett House,GBREX002
+Peterhead,GB000416
+Plymouth,GB000250
+Plymouth Docks,GB003050
+Poole Ferry Terminal,GB000291
+Portbury,GB005190
+Portree (Isle Of Skye),GB003240
+Portsmouth,GB000292
+Prestwick Airport,GB000070
+Purfleet Thames Terminal,GB000149
+Ramsgate,GB000047
+Reading,GB000155
+Ridham c/o Simps Team,GB005110
+Rochester c/o Simps Team,GB005120
+Rosyth,GB000057
+Salford National Clearance Hub,GB000383
+Scotland Frontier,GB000067
+Scotland Inland Auth. Cee/Cor,GB000063
+Scrabster,GB000062
+Sheerness,GB000045
+Shoreham c/o Simps Team,GB005130
+South Shields c/o Simps Team,GB005140
+Southampton,GB000290
+Southampton (Eastleigh) Airport,GB003090
+Southend Airport,GB000091
+Southend Direct Exports,GBSOU001
+Stansted Airport FCT,GB000121
+Stornoway (Isle Of Lewis),GB003260
+Stornoway Airport,GB003270
+Swansea,GB000123
+Teesport,GB000219
+Thames Europort (DIFT),GB000099
+Thamesport,GB000150
+Tilbury,GB000093
+Uk North Auth Consignor/Nees,GB000081
+UK South Auth Consignor/Nees,GB000246
+UKBA East Midlands Airport,GB000245
+UKBF Bristol International Airport,GB000029
+Warrenpoint,GB005160
+Workington,GB003280
