+++ test/services/PackageTypesSpec.scala
@@ -19,19 +19,21 @@ package services
 import services.model.PackageType
 import uk.gov.hmrc.play.test.UnitSpec
 
-class PackageTypeSpec extends UnitSpec {
+class PackageTypesSpec extends UnitSpec {
 
   "Package type list" should {
+    
     "return package types containing commas and quotes" in {
-      val somePackageTypes: List[PackageType] =
-        PackageTypes.all.filter((packageType: PackageType) => packageType.code == "43")
+      val somePackageTypes = PackageTypes.all.filter(_.code == "43")
+      
       somePackageTypes shouldBe List(PackageType("43", "Bag, super bulk"))
     }
 
     "return package types' with codes in alphabetical order of name" in {
       val expectedCodes = Set("43", "AD", "ZZ")
-      val somePackageTypes: List[PackageType] =
-        PackageTypes.all.filter((packageType: PackageType) => expectedCodes.contains(packageType.code))
+      
+      val somePackageTypes = PackageTypes.all.filter(packageType => expectedCodes.contains(packageType.code))
+      
       somePackageTypes shouldBe List(
         PackageType("43", "Bag, super bulk"),
         PackageType("ZZ", "Defined mutually"),
